+++ kfinance/tool_calling/__init__.py
@@ -22,7 +22,7 @@ from kfinance.tool_calling.get_isin_from_ticker import GetIsinFromTicker
 from kfinance.tool_calling.get_latest import GetLatest
 from kfinance.tool_calling.get_n_quarters_ago import GetNQuartersAgo
 from kfinance.tool_calling.get_prices_from_identifier import GetPricesFromIdentifier
-from kfinance.tool_calling.resolve_identifiers import ResolveIdentifiers
+from kfinance.tool_calling.resolve_identifier import ResolveIdentifier
 from kfinance.tool_calling.shared_models import KfinanceTool
 
 
@@ -39,5 +39,5 @@ ALL_TOOLS: list[Type[KfinanceTool]] = [
     GetFinancialStatementFromIdentifier,
     GetFinancialLineItemFromIdentifier,
     GetBusinessRelationshipFromIdentifier,
-    ResolveIdentifiers,
+    ResolveIdentifier,
 ]